Enhance dep-relations for paasng.platform.region, more
